Split visit buttons from project details component
